Refactor flagging system
The `block_set/unset_flags` functions now take an `OmniBlock` instead of
an `OmniBlockStatusFlags` as the first argument. This is for consistency
with the respective sample flagging functions, and to implement some
extra functionality to control sample validity when changing block
flags.

This also required the creation of specific flagging functions for
`OmniMetaBlock`.